model var type

+++ repo/store/shortlink/model.go
@@ -4,7 +4,7 @@ import "time"
 
 // Short4LongLinkModel 短链数据库model
 type Short4LongLinkModel struct {
-	Id        int       `xorm:"pk autoincr" json:"id"`
+	Id        int64     `xorm:"pk autoincr" json:"id"`
 	ShortLink string    `xorm:"index" json:"shortLink"`
 	LongLink  string    `json:"longLink"`
 	Ctime     time.Time `xorm:"created" json:"ctime"`
